added condition isAdmin

+++ src/app/api/categories/route.js
@@ -1,26 +1,34 @@
 import { Category } from "@/models/Category"
 import mongoose from "mongoose"
+import { isAdmin } from "../auth/[...nextauth]/route"
 
 export async function POST(req) {
    mongoose.connect(process.env.MONGO_URL)
    const { name } = await req.json()
-   const categoryDoc = await Category.create({ name })
-   return Response.json(categoryDoc)
+   if (await isAdmin()) {
+    const categoryDoc = await Category.create({name});
+    return Response.json(categoryDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function DELETE(req) {
     mongoose.connect(process.env.MONGO_URL)
     const { _id, name } = await req.json()
-    await Category.deleteOne({_id}, {name})
+    if (await isAdmin()) {
+        await Category.deleteOne({_id}, {name})
+    }
     return Response.json(true)
 }
 
 export async function PUT(req) {
     mongoose.connect(process.env.MONGO_URL)
     const { _id, name } = await req.json()
-    await Category.updateOne({_id}, {name})
-    return Response.json(true)
-
+    if (await isAdmin()) {
+        await Category.updateOne({_id}, {name});
+    }
+    return Response.json(true);
 }
 
 export async function GET() {
